+++ apps/imlib/src/imboy_req.erl
@@ -24,7 +24,7 @@ page_size(Req) ->
     {Size2, _} = string:to_integer(Size),
     pase_page_size(Page2, Size2).
 
-
+% imboy_req:get_int()
 get_int(Key, Req, Def) ->
     #{Key := Val} = cowboy_req:match_qs([{Key, [], Def}], Req),
     if